Extract archive operations from agent manager into dedicated module
Move beginArchive, executeArchive, and deleteAgentDirect (~270 lines)
from the 1734-line manager.ts into agents/archive.ts, reducing the
manager to 1463 lines. The public API is unchanged — manager methods
now delegate to the extracted functions via a narrow ArchiveDeps
interface.
